COC-62-feat(#89): 스페이스 디테일 Handler 분리
